% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr.R
\name{flowr}
\alias{flowr}
\title{Returns a graph}
\usage{
flowr(tib_elems, element = NA, layout = "kk", highlighted = NA)
}
\arguments{
\item{layout}{the type of layout: either "kk" (for a flower-like representation, the default) or "sugiyama" (for a tree-like representation)}

\item{highlighted}{functions or arguments that are to be highlighted in the graph}
}
\value{
a graph
}
\description{
Returns a graph
}
\examples{
library(flowrpowr)
tib_elems=tibble(
  root="gwek",
  pack_or_fun="package",
  elems=c("pouet_boum.bam",
          "pouet_hop_bam",
          "hop_boum.bam",
          "hop_boum_pouet_tac",
          "hop_boum_pouet_tocBim",
          "hop_boum_pouet_tocBimPaf",
          "hop_boum_pouet_tocBoum"))
 flowr(tib_elems)
}
