% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr_package.R
\name{flowr_package}
\alias{flowr_package}
\title{Returns tibble with frequencies of cat (lemma or word for instance) in data}
\usage{
flowr_package(packagename, element = "")
}
\arguments{
\item{data}{a tibble}

\item{cat}{words or lemmas, for instance}

\item{criterion}{should the output be filtered by top specificities by category ('top_n') or according to a minimum value of specificity ("spec_min").}

\item{top_n}{in case criterion=='top_n', how many items by category should be kept (defaults to 50)}

\item{spec_min}{in case criterion=='min_spec', which is the minimum specificity for an item to be kept (defaults to 2)}
}
\value{
a tibble of cat frequencies
}
\description{
Returns tibble with frequencies of cat (lemma or word for instance) in data
}
\examples{
library(flowrpowr)
flowr_package("tidygraph")
flowr_package(packagename="tidygraph",
              element="edge")
}
