% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regather_false_pairs.R
\name{regather_false_pairs}
\alias{regather_false_pairs}
\title{When one source node is linked only to one target node they are considered a false pair}
\usage{
regather_false_pairs(tib)
}
\description{
When one source node is linked only to one target node they are considered a false pair
}
