% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr.R
\name{flowr}
\alias{flowr}
\title{Returns a graph}
\usage{
flowr(list_elems, layout = "kk")
}
\arguments{
\item{layout}{the type of layout: either "kk" (for a flower-like representation, the default) or "sugiyama" (for a tree-like representation)}
}
\value{
a graph
}
\description{
Returns a graph
}
\examples{
library(flowrpowr)
flowr(list_elems=tibble(
  root="gwek",
  pack_or_fun="package",
  elems=c("pouet_boum.bam",
          "pouet_hop_bam",
          "hop_boum.bam",
          "hop_boum_pouet_tac",
          "hop_boum_pouet_toc")))
}
