% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr_function.R
\name{flowr_function}
\alias{flowr_function}
\title{Returns a graph showing all the arguments of a function or the arguments of a function with a given element as part of their name}
\usage{
flowr_function(functionnames, element = NA, layout = "kk")
}
\arguments{
\item{element}{if provided, part of the name of the argument of interest}

\item{layout}{the type of layout: either "kk" (for a flower-like representation, the default) or "sugiyama" (for a tree-like representation)}

\item{functionname}{the name of a function in an installed and loaded package}
}
\value{
a graph
}
\description{
Returns a graph showing all the arguments of a function or the arguments of a function with a given element as part of their name
}
\examples{
library(flowrpowr)
flowr_function("geom_edge_link")
flowr_function("geom_edge_link",
               element="label")
flowr_function(c("geom_violin","geom_point","geom_boxplot"))
flowr_function(c("geom_violin","geom_point","geom_boxplot"),layout="sugiyama")
}
