% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr_package.R
\name{flowr_package}
\alias{flowr_package}
\title{Returns a graph showing all the arguments of a function or the arguments of a function with a given element as part of their name}
\usage{
flowr_package(
  packagename,
  layout = "kk",
  element = NA,
  highlighted = NA,
  return_tibble = FALSE
)
}
\arguments{
\item{layout}{the type of layout: either "kk" (for a flower-like representation, the default) or "sugiyama" (for a tree-like representation)}

\item{element}{if provided, part of the name of the argument of interest}

\item{functionname}{the name of an installed and loaded package}
}
\value{
a graph
}
\description{
Returns a graph showing all the arguments of a function or the arguments of a function with a given element as part of their name
}
\examples{
library(flowrpowr)
flowr_package("tidygraph")
flowr_package(packagename="tidygraph",
              element="node", layout="sugiyama")
flowr_package(packagename="tidygraph",
              element="node_rank")
}
